yesterday_listener/requirements.txt
requests==2.19.1
google-cloud-pubsub==0.40.0
dnspython==1.16.0
pymongo==3.8.0